surprise-vs-angry/requirements.txt
tensorflow
keras
numpy
matplotlib
Pillow